+++ src/extension.ts
@@ -1,73 +1,88 @@
 import * as vscode from "vscode";
-
-interface TabInfo {
-  lastActive: number;
-}
-
-const tabMap: Map<string, TabInfo> = new Map();
-
-const MAX_TABS = 10;
-const INACTIVE_TIMEOUT = 300000; // 5 minutes
-const INACTIVE_CHECK_INTERVAL = 5000; // 5 seconds
+import { getTabMap, getOpenEditorTabs, getLRUClosableTabs } from "./utils/tabs";
+import { toggleCommands } from "./utils/constants";
+import {
+  DEFAULT_INACTIVE_CHECK_INTERVAL,
+  DEFAULT_INACTIVE_TIMEOUT,
+  DEFAULT_MAX_TABS,
+} from "./utils/constants";
 
 export function activate(context: vscode.ExtensionContext) {
+  const tabMap = getTabMap();
   const config = vscode.workspace.getConfiguration("tabCloser");
-  let enabled = config.get<boolean>("enabled", true);
-  let inactiveTimeout = config.get<number>("inactiveTimeout", INACTIVE_TIMEOUT);
-  let maxTabs = config.get<number>("maxTabs", MAX_TABS);
+  const enabled: { [x in string]: boolean } = {
+    global: config.get<boolean>("enabled", true),
+    inactiveTimeout: config.get<boolean>("inactiveTimeoutEnabled", true),
+    maxTabs: config.get<boolean>("maxTabsEnabled", true),
+  };
 
-  const toggleCommand = vscode.commands.registerCommand(
-    "tabCloser.toggle",
-    async () => {
+  let inactiveTimeout = config.get<number>(
+    "inactiveTimeout",
+    DEFAULT_INACTIVE_TIMEOUT
+  );
+  let maxTabs = config.get<number>("maxTabs", DEFAULT_MAX_TABS);
+
+  const toggleEnabledCommand = ({
+    command,
+    configKey,
+    key,
+  }: {
+    command: string;
+    configKey: string;
+    key: string;
+  }) => {
+    const toggleCommand = vscode.commands.registerCommand(command, async () => {
       const config = vscode.workspace.getConfiguration("tabCloser");
-      const current = enabled ?? false;
+      const current = enabled[key] ?? false;
       await config.update(
-        "enabled",
+        configKey,
         !current,
         vscode.ConfigurationTarget.Global
       );
-      enabled = !current;
-      vscode.window.showInformationMessage(
-        `Tab Closer is now ${enabled ? "enabled" : "disabled"}.`
-      );
-    }
-  );
-  context.subscriptions.push(toggleCommand);
+    });
+    context.subscriptions.push(toggleCommand);
+  };
+  toggleCommands.forEach((command) => {
+    toggleEnabledCommand(command);
+  });
+
+  // Initialize the tabMap with currently open tabs
+  getOpenEditorTabs().forEach((tab) => {
+    const uri = (tab.input as vscode.TabInputText).uri.toString();
+    tabMap.set(uri, { lastActive: Date.now() });
+  });
 
   vscode.workspace.onDidChangeConfiguration((e) => {
     if (
       e.affectsConfiguration("tabCloser.enabled") ||
+      e.affectsConfiguration("tabCloser.inactiveTimeoutEnabled") ||
+      e.affectsConfiguration("tabCloser.maxTabsEnabled") ||
       e.affectsConfiguration("tabCloser.inactiveTimeout") ||
       e.affectsConfiguration("tabCloser.maxTabs")
     ) {
       const updatedConfig = vscode.workspace.getConfiguration("tabCloser");
-      enabled = updatedConfig.get<boolean>("enabled", true);
+      enabled.global = updatedConfig.get<boolean>("enabled", true);
+      enabled.inactiveTimeout = updatedConfig.get<boolean>(
+        "inactiveTimeoutEnabled",
+        true
+      );
+      enabled.maxTabs = updatedConfig.get<boolean>("maxTabsEnabled", true);
       inactiveTimeout = updatedConfig.get<number>(
         "inactiveTimeout",
-        INACTIVE_TIMEOUT
+        DEFAULT_INACTIVE_TIMEOUT
       );
-      maxTabs = updatedConfig.get<number>("maxTabs", MAX_TABS);
+      maxTabs = updatedConfig.get<number>("maxTabs", DEFAULT_MAX_TABS);
+      closeInactiveTabs();
+      closeLeastRecentlyUsedTabs();
     }
   });
 
-  vscode.window.onDidChangeActiveTextEditor((editor) => {
-    const doc = editor?.document;
-    if (doc && doc.uri.scheme === "file") {
-      tabMap.set(doc.uri.toString(), { lastActive: Date.now() });
-    }
-  });
-
-  vscode.workspace.onDidCloseTextDocument((doc) => {
-    tabMap.delete(doc.uri.toString());
-  });
-
-  const interval = setInterval(() => {
-    if (!enabled) {
+  // tab closing handlers
+  function closeInactiveTabs() {
+    if (!enabled.global || !enabled.inactiveTimeout) {
       return;
     }
-
     const now = Date.now();
-
     for (const [uri, info] of tabMap.entries()) {
       if (now - info.lastActive > inactiveTimeout) {
         const editor = vscode.window.visibleTextEditors.find(
@@ -80,15 +95,14 @@ export function activate(context: vscode.ExtensionContext) {
         }
       }
     }
+  }
+  function closeLeastRecentlyUsedTabs() {
+    if (!enabled.global || !enabled.maxTabs) {
+      return;
+    }
 
-    const openTabs = getOpenEditorTabs().map(
-      (tab) => tab.input as vscode.TabInputText
-    );
     const closableTabs = getLRUClosableTabs();
 
-    vscode.window.showInformationMessage(
-      `Open tabs: ${openTabs.length}, Closable tabs: ${closableTabs.length}, Max tabs: ${maxTabs}`
-    );
     while (tabMap.size > maxTabs && closableTabs.length > 0) {
       const { tab } = closableTabs.shift()!;
       vscode.window.tabGroups.close(tab);
@@ -98,50 +112,41 @@ export function activate(context: vscode.ExtensionContext) {
         ).input.uri.toString()
       );
     }
-  }, INACTIVE_CHECK_INTERVAL);
+  }
+  vscode.commands.registerCommand(
+    "tabCloser.closeInactiveTabs",
+    closeInactiveTabs
+  );
+  vscode.commands.registerCommand(
+    "tabCloser.closeLeastRecentlyUsedTabs",
+    closeLeastRecentlyUsedTabs
+  );
+
+  vscode.window.onDidChangeActiveTextEditor((editor) => {
+    const doc = editor?.document;
+    if (doc && doc.uri.scheme === "file") {
+      const isNew = !tabMap.has(doc.uri.toString());
+      tabMap.set(doc.uri.toString(), { lastActive: Date.now() });
+      if (isNew && enabled.maxTabs) {
+        closeLeastRecentlyUsedTabs();
+      }
+    }
+  });
+
+  vscode.workspace.onDidCloseTextDocument((doc) => {
+    tabMap.delete(doc.uri.toString());
+  });
+
+  const interval = setInterval(() => {
+    if (!enabled.global) {
+      return;
+    }
+    if (enabled.inactiveTimeout) {
+      closeInactiveTabs();
+    }
+  }, DEFAULT_INACTIVE_CHECK_INTERVAL);
 
   context.subscriptions.push({ dispose: () => clearInterval(interval) });
-}
-
-function getOpenEditorTabs(): vscode.Tab[] {
-  return vscode.window.tabGroups.all
-    .flatMap((group) => group.tabs)
-    .filter(
-      (tab) =>
-        tab.input instanceof vscode.TabInputText &&
-        tab.input.uri.scheme === "file"
-    );
-}
-
-function getClosableTabs(): vscode.Tab[] {
-  return vscode.window.tabGroups.all
-    .flatMap((group) => group.tabs)
-    .filter(
-      (tab): tab is vscode.Tab & { input: vscode.TabInputText } =>
-        tab.input instanceof vscode.TabInputText &&
-        tab.input.uri.scheme === "file" &&
-        !vscode.workspace.textDocuments.find(
-          (d) =>
-            d.uri.toString() ===
-            (
-              tab as vscode.Tab & { input: vscode.TabInputText }
-            ).input.uri.toString()
-        )?.isDirty
-    );
-}
-
-function getLRUClosableTabs(): { tab: vscode.Tab; lastActive: number }[] {
-  return getClosableTabs()
-    .map((tab) => ({
-      tab,
-      lastActive:
-        tabMap.get(
-          (
-            tab as vscode.Tab & { input: vscode.TabInputText }
-          ).input.uri.toString()
-        )?.lastActive ?? 0,
-    }))
-    .sort((a, b) => a.lastActive - b.lastActive); // Oldest first
 }
 
 export function deactivate() {}