+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## [0.2.0]
+
+### Added
+
+- Toggle command to enable/disable individual features
+- Commands to manually run `closeInactiveTabs` and `closeLeastRecentlyUsed` tab actions
+
+### Fixed
+
+- Fixed bug where the initial tab was not reflected in the list of current tabs
+
 ## [0.1.0]
 
 ### Added